+++ code/src/cljs/sixsq/nuvla/ui/profile/views.cljs
@@ -1128,7 +1128,6 @@
                     currency email] :or {balance 0}} @customer-info
             {:keys [street-address city country
                     postal-code]} address
-
             balance (* balance -1)]
         [ui/Segment {:padded  true
                      :color   "grey"
@@ -1143,10 +1142,9 @@
            [ui/TableRow
             [ui/TableCell [:b (@tr [:street-address])]]
             [ui/TableCell street-address [:br] postal-code " - " city [:br] country]]
-           (when email
-             [ui/TableRow
-              [ui/TableCell [:b "Email"]]
-              [ui/TableCell email]])
+           [ui/TableRow
+            [ui/TableCell [:b (str/capitalize (@tr [:email]))]]
+            [ui/TableCell (or email "-")]]
            [ui/TableRow {:negative (neg? balance)
                          :positive (pos? balance)}
             [ui/TableCell {:width 5} [:b (str/capitalize (@tr [:balance]))]]