+++ CHANGELOG.md
@@ -4,8 +4,8 @@
 
 ### Added
 
-- Profile - Display customer email
 - Profile - Group tab show new segment with group hierarchy
+- Profile - Display customer email
 
 ### Changed